requirements.txt
djangoframework
stripe
